// src/services/theme/interfaces/IThemeColors.ts
export interface ThemeColors {
  primaryColor: string;
  primaryLight: string;
  secondaryColor: string;
  bgColor: string;
  textColor: string;
  lightGray: string;
  mediumGray: string;
  darkGray: string;
  successColor: string;
  errorColor: string;
  infoColor: string;
  userMessageBg: string;
  aiMessageBg: string;
}